requirements.txt
# Minimal serving dependencies
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.3.2
PyYAML==6.0.1

# MLflow for model loading
mlflow==2.11.3
sqlalchemy

# Prefect for orchestration
prefect==2.14.15

# FastAPI for serving
fastapi==0.95.2
uvicorn==0.35.0
pydantic==1.10.13
typing-extensions==4.14.1

# Monitoring
evidently==0.3.2
griffe==0.20.0

ruff=0.12.4
black=25.1.0